First practical usage of the OpenGL framework built.
