+++ Space_Invaders.vcxproj.filters
@@ -177,6 +177,12 @@
     <ClInclude Include="dep\include\space_invaders.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="dep\include\Entitys\entity.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="dep\include\Entitys\ufo.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Image Include="res\sprite_sheet.jpg">
