+++ builds/0_05- menu_showcase/res/score.dat
@@ -1,0 +1,1 @@
+0
