+++ builds/0_05- menu_showcase/res/shaders/bitmap_frag.shader
@@ -1,0 +1,15 @@
+#version 330 core
+
+in vec2 textCoord;
+in vec4 vertexColor;
+
+uniform sampler2D myTexture;
+
+out vec4 fragColor;
+
+void main() {
+	vec4 textColor = texture(myTexture, textCoord) * vertexColor;
+	if (textColor.a < 0.1)
+		discard;
+	fragColor = textColor;
+}
