+++ dep/include/Entitys/ufo.h
@@ -1,0 +1,13 @@
+#ifndef UFO_H
+#define UFO_H
+
+#include "Entitys/entity.h"
+
+class UFO : public Entity {
+public:
+
+private:
+
+};
+
+#endif
